requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.40
pydantic==2.11.4
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pytest==8.3.5
pytest-cov==6.1.1
pytest-asyncio==0.23.6
alembic==1.15.2
asyncpg==0.29.0
aiosqlite==0.20.0
email-validator==2.1.1
